+++ app/controllers/concerns/exception_handler.rb
@@ -7,7 +7,14 @@ module ExceptionHandler
     end
 
     rescue_from Kinedu::Exceptions::MissingToken do |e|
-      render json: { errors: ["Token not found"] }, status: :bad_request
+      render json: { errors: I18n.t('controller.application_controller.missing_token') }, status: :bad_request
+    end
+
+    rescue_from Kinedu::Exceptions::EndTimeError do |e|
+      render json: {
+        errors: [ I18n.t('controller.application_controller.end_time_error')]
+        },
+        status: :bad_request
     end
   end
 end