+++ app/lib/kinedu/exceptions.rb
@@ -1,5 +1,6 @@
 module Kinedu::Exceptions
   class MissingToken < StandardError; end
+  class EndTimeError < StandardError; end
   class Authentication < StandardError 
     def initialize(msg)
       super(msg)